content-repurposing-agent/requirements.txt
# Core Agent Framework
phidata>=2.0.0

# LLM Providers
anthropic>=0.25.0
google-generativeai>=0.4.0
openai>=1.12.0

# UI
streamlit>=1.31.0

# Image Generation
replicate>=0.25.0  # For Stable Diffusion alternative
# openai  # Already included above for DALL-E

# Social Media APIs
tweepy>=4.14.0  # Twitter API
requests>=2.31.0  # For LinkedIn and other HTTP APIs
instagrapi>=2.0.0  # Instagram API

# SEO & Research Tools
serpapi>=0.1.5  # Google Search API wrapper
# Alternative: requests for direct API calls

# Database & Storage
sqlalchemy>=2.0.25
python-dotenv>=1.0.0

# Utilities
pillow>=10.2.0  # Image processing
python-dateutil>=2.8.2
pydantic>=2.6.0  # Data validation
